Changed to FSM, removed keyboard
This is a big refactory of the code, but it was hugly needed. The main problem
so far was that each action was a function, therefore, adding actions
translated into adding functions. At some point, the functions had to control
too many things:

take -> walk to the object
	look at it
	animate the arm rising
	put it in the inventory
	animate the arm going down

Writing the commands (walk, take, look) in such a way that they did not interfere
with each other was growing too much in complexity, also, it was hard to read. So,
I changed to the following:

Each character (so far Cole, the only character that exists) has a queue of
actions. Each command (walk, take, look, ...) gets translated into a set of
actions. The character constantly looks at its queue, and goes making the
actions one by one. In this way, take gets literally translated to the actions

take -> walk, face(object), animation(rise_arm), take(), animate(lower_arm)

This will make much much easier to write commands, and to add more characters
to the game. In the way everything is writen, you can literally control two or
more characters (of course, giving orders one at the time!).