Added inventory
The inventory is a part of the GUI. It is assigned to the player
by the level, in this way, we will be able to have one inventory
per player. The player interacts by "adding" things to the
inventory. This adds an image to the GUI, in the inventory sector.
No further interaction is coded so far.