Separated actions into their own file
Actions keep growing in quantity and complexity. We went from "walk"
to "take", "read", "examine" and now "use". "Use" is not simply
another action, because it need another object to be performed:
Use A with B. I decided to make a class for actions, in this way
they are already defined for the whole game. Otherwise, in the
level we had 1 array and 2 dictionaries.. that was not going to
be easy to replicate in further levels.