Inventory items can be used
Actions are now either "immediate", "to_combine" or "combined". "To_combine"
are actions that need to be combined with an object in order to be
trigered. The only function like this is "use". "Use" needs to be combined
with an object... Use object with <something else>. This is now implemented.