Project Refactory
Created a node for "Point & Click", which handles the whole logic
of the point (looking stuff with the mouse) and click (executing
an action). I have refactored the rest of the project.
Now, the Level "House" is not attached to the "Point and Click"
system itself. This is because we will want to use the same Point
and Click in different levels.

In the future, we will make Cole go out of the house. There, we will
want to mantain our point and click, without having to write it
again.

At the same time, we want to be able to make the "Level" control
only level variables. The level should only care about if some
conditions within itself have been met. Not about how to point and
click.