Added the action "Open" and interaction with the wardrobe
This commit adds interaction with the wardrobe of the left room.
Now we can open it, and a Tween will animate its movement.

Also, a problem with the "face_direction" has been fixed. Now
Cole checks where the object is respect to him and the current
camera, and then rotates accordingly.