Changed WalkTo state to WalkPath
The FSM state WalkTo was computing the path to walk. This imposes
the player to have a "navigation" property. Now the player
computes the path, and the FSM simply makes it walk. There is still
the "face to", which forces the player to have the "face to" method.
Will think on how to remove it in further commits.