Changed FSM and Constrained Actions
Now, an object has a set of actions that can be performed on it. The
"point" shows only those actions available when hovering over the object.

The FSM also changed, I introduced the state "PerformActionOnObject".
This state allows to queue a function call to an object. By doing so,
we can simplify some actions as "open" and "use".

Interaction with lights has been implemented to showcase the new "use"
vs "use_item". One is an immediate action, that allows to switch the
light... vs "using an item" where we need to combine objects.