Changed how transition works
I rewrote how transitions work. For this I needed to further change
how actions work on objects. Now, the objects receive as a parameter
who did the action on them. This actually makes sense if we think
about it. We could, for example, want to "move" something only if
the "strongest" player is doing the action. Or "lockpick" a door
only if the "thief" character is doing so.