Reworked classes to better separate P&C logic
The Point and Click classes were renamed, and some logic was changed,
for example, FSM was divided into STATES and Queueu.

The point and click system only knows that there is a player and
objects. When something is clicked: it eithers combine the action, if
it is combinable, or tells the player to do_action_on_object.

Now Cole is an instance of the new class Player. A player is
basically an object with a queue, that process the States in the
Queue (process_queue) and has the function "do_action_on_object".
"do_action_on_object" translates the Action into "call the function
Action.function in myself".

When a player wants to make an action, it actually creates a STATE
and adds it to its queue. Then, the STATE will eventually call more
functions from the player, in a timely way.