Simplified CutScenes through the use of signals
Now, a player "signals" each time it finished an action. This is
done through a new state called "Finished". For example:
walk_to -> animate(walk), move_player, animate(idle), finished.

In this way, we can know when an action has finished, meaning
that we can move to the next one in the cutscene.