Simplified 3D models for winning performance
The scene so far was a bunch of 3D objects, all manually placed. Now,
I have passed all the "magica voxel" files through Blender, merged
them and imported them in Godot as a single object. The performance
will is clear since the previous version was not able to run smoothly
in my notebook, and now it is.